feat: settings module (#83)
