+++ core/common/src/main/java/jp/mydns/kokoichi0206/common/BuildConfigWrapper.kt
@@ -5,4 +5,5 @@ package jp.mydns.kokoichi0206.common
  */
 data class BuildConfigWrapper(
     val API_KEY: String,
+    val VERSION: String,
 )
