+++ core/common/src/main/java/jp/mydns/kokoichi0206/common/FunctionUtils.kt
@@ -1,5 +1,9 @@
 package jp.mydns.kokoichi0206.common
 
+import androidx.compose.ui.graphics.Color
+import jp.mydns.kokoichi0206.common.ui.theme.BaseColorH
+import jp.mydns.kokoichi0206.common.ui.theme.BaseColorN
+import jp.mydns.kokoichi0206.common.ui.theme.BaseColorS
 import java.time.LocalTime
 
 
@@ -30,5 +34,49 @@ fun getGenerationLooper(group: String): List<String> {
             Constants.POSSIBLE_GENERATIONS_H
         else ->
             Constants.POSSIBLE_GENERATIONS_N
+    }
+}
+
+/**
+ * Get baseColor of the group by group name (NOGIZAKA or 乃木坂)
+ *
+ * @param group Group name
+ * @return BaseColor of the passed group (Color type)
+ */
+fun getBaseColor(group: String): Color {
+
+    return when (group) {
+        GroupName.NOGIZAKA.name, GroupName.NOGIZAKA.jname -> {
+            BaseColorN
+        }
+        GroupName.HINATAZAKA.name, GroupName.HINATAZAKA.jname -> {
+            BaseColorH
+        }
+        GroupName.SAKURAZAKA.name, GroupName.SAKURAZAKA.jname -> {
+            BaseColorS
+        }
+        else -> Color.Black
+    }
+}
+
+/**
+ * Get subColor of the group by group name (NOGIZAKA or 乃木坂)
+ *
+ * @param group Group name
+ * @return SubColor of the passed group (Color type)
+ */
+fun getSubColor(group: String): Color {
+
+    return when (group) {
+        GroupName.NOGIZAKA.name, GroupName.NOGIZAKA.jname -> {
+            BaseColorN
+        }
+        GroupName.HINATAZAKA.name, GroupName.HINATAZAKA.jname -> {
+            BaseColorH
+        }
+        GroupName.SAKURAZAKA.name, GroupName.SAKURAZAKA.jname -> {
+            BaseColorS
+        }
+        else -> BaseColorN
     }
 }
