+++ feature/settings/build.gradle
@@ -1,0 +1,63 @@
+plugins {
+    id 'com.android.library'
+    id 'org.jetbrains.kotlin.android'
+    id 'kotlin-kapt'
+}
+
+android {
+    namespace 'jp.mydns.kokoichi0206.feature.settings'
+    compileSdk build_versions.target_sdk
+
+    defaultConfig {
+        minSdk build_versions.min_sdk
+    }
+    composeOptions {
+        kotlinCompilerExtensionVersion build_versions.compose
+    }
+    buildFeatures {
+        compose true
+    }
+}
+
+dependencies {
+    implementation deps.compose.ui
+    implementation deps.material.compose
+    implementation deps.material.material
+
+    implementation deps.lifecycle.runtime
+
+    // hilt
+    implementation deps.hilt.navigation
+    kapt deps.hilt.compiler
+
+    // Compose dependencies
+    implementation deps.compose.lifecycle_viewmodel
+    implementation deps.compose.navigation
+    implementation deps.accompanist.flowlayout
+
+    // Coroutines
+    implementation deps.coroutines.core
+    implementation deps.coroutines.android
+
+    // Coroutine Lifecycle Scopes
+    implementation deps.lifecycle.viewmodel_ktx
+
+    // Coil
+    implementation deps.coil
+
+    // Inject
+    implementation deps.hilt.android
+
+    // Animated navigation
+    implementation deps.accompanist.animation
+
+    // DataStore
+    implementation deps.datastore.preferences
+    // optional - RxJava3 support
+    implementation deps.datastore.rxjava3
+
+    // modules
+    implementation project(":core:common")
+    implementation project(":core:model")
+    implementation project(":core:domain")
+}
