+++ feature/settings/src/main/java/jp/mydns/kokoichi0206/settings/TestTags.kt
@@ -1,0 +1,49 @@
+package jp.mydns.kokoichi0206.settings
+
+object TestTags {
+    // Setting Screen
+    const val SETTING_TITLE = "SETTING_TITLE"
+    const val SETTING_VERSION = "SETTING_VERSION"
+    const val SNACK_BAR_TEXT = "SNACK_BAR_TEXT"
+
+    const val CACHE_CLEAR_DIALOG_TITLE = "CACHE_CLEAR_DIALOG_TITLE"
+    const val CACHE_CLEAR_DIALOG_BODY = "CACHE_CLEAR_DIALOG_BODY"
+    const val CACHE_CLEAR_DIALOG_OK = "CACHE_CLEAR_DIALOG_OK"
+    const val CACHE_CLEAR_DIALOG_CANCEL = "CACHE_CLEAR_DIALOG_CANCEL"
+
+    // UpdateBlog Screen
+    const val UPDATE_BLOG_TITLE = "UPDATE_BLOG_TITLE"
+    const val UPDATE_BLOG_BODY = "UPDATE_BLOG_BODY"
+    const val UPDATE_BLOG_OK_BUTTON = "UPDATE_BLOG_OK_BUTTON"
+    const val UPDATE_BLOG_CANCEL_BUTTON = "UPDATE_BLOG_CANCEL_BUTTON"
+    const val UPDATE_BLOG_DIALOG_TITLE = "UPDATE_BLOG_DIALOG_TITLE"
+    const val UPDATE_BLOG_DIALOG_BODY = "UPDATE_BLOG_DIALOG_BODY"
+    const val UPDATE_BLOG_DIALOG_OK_BUTTON = "UPDATE_BLOG_DIALOG_OK_BUTTON"
+
+    // ReportIssue Screen
+    const val REPORT_ISSUE_TITLE = "REPORT_ISSUE_TITLE"
+    const val REPORT_ISSUE_BODY = "REPORT_ISSUE_BODY"
+    const val REPORT_ISSUE_BUTTON = "REPORT_ISSUE_BUTTON"
+    const val REPORT_ISSUE_DIALOG_TITLE = "REPORT_ISSUE_DIALOG_TITLE"
+    const val REPORT_ISSUE_DIALOG_BODY = "REPORT_ISSUE_DIALOG_BODY"
+    const val REPORT_ISSUE_DIALOG_OK = "REPORT_ISSUE_DIALOG_OK"
+    const val REPORT_ISSUE_DIALOG_CANCEL = "REPORT_ISSUE_DIALOG_CANCEL"
+
+    // SetTheme Screen
+    const val SET_THEME_BACK_ARROW = "SET_THEME_BACK_ARROW"
+    const val SET_THEME_THEME = "SET_THEME_THEME"
+
+    // ShareApp Screen
+    const val SHARE_APP_MESSAGE = "SHARE_APP_MESSAGE"
+    const val SHARE_APP_QR_CODE = "SHARE_APP_QR_CODE"
+
+    // AboutApp Screen
+    const val ABOUT_APP_APP_ICON = "ABOUT_APP_APP_ICON"
+    const val ABOUT_APP_VERSION_TITLE = "ABOUT_APP_VERSION_TITLE"
+    const val ABOUT_APP_VERSION_CONTENT = "ABOUT_APP_VERSION_CONTENT"
+    const val ABOUT_APP_USERID_TITLE = "ABOUT_APP_USERID_TITLE"
+    const val ABOUT_APP_USERID_CONTENT = "ABOUT_APP_USERID_CONTENT"
+
+    // Setting utils
+    const val SETTING_TOP_BAR = "SETTING_TOP_BAR"
+}
