+++ feature/settings/src/main/java/jp/mydns/kokoichi0206/settings/components/SettingTopBar.kt
@@ -1,4 +1,4 @@
-package jp.mydns.kokoichi0206.sakamichiapp.presentation.setting.components
+package jp.mydns.kokoichi0206.settings.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -18,13 +18,13 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
-import jp.mydns.kokoichi0206.sakamichiapp.R
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.setting.ThemeType
+import jp.mydns.kokoichi0206.settings.ThemeType
 import jp.mydns.kokoichi0206.common.ui.theme.IconSizeMedium
 import jp.mydns.kokoichi0206.common.ui.theme.SpaceMedium
 import jp.mydns.kokoichi0206.common.ui.theme.SpaceTiny
 import jp.mydns.kokoichi0206.common.ui.theme.Typography
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.util.TestTags
+import jp.mydns.kokoichi0206.feature.settings.R
+import jp.mydns.kokoichi0206.settings.TestTags
 
 @Composable
 fun SettingTopBar(
