+++ feature/settings/src/main/java/jp/mydns/kokoichi0206/settings/pages/ReportIssueScreen.kt
@@ -1,4 +1,4 @@
-package jp.mydns.kokoichi0206.sakamichiapp.presentation.setting.pages
+package jp.mydns.kokoichi0206.settings.pages
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -17,13 +17,13 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import jp.mydns.kokoichi0206.common.Constants
-import jp.mydns.kokoichi0206.sakamichiapp.R
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.setting.SettingsUiState
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.setting.SettingsViewModel
+import jp.mydns.kokoichi0206.feature.settings.R
+import jp.mydns.kokoichi0206.settings.SettingsUiState
+import jp.mydns.kokoichi0206.settings.SettingsViewModel
 import jp.mydns.kokoichi0206.common.ui.theme.SpaceMedium
 import jp.mydns.kokoichi0206.common.ui.theme.SpaceSmall
 import jp.mydns.kokoichi0206.common.ui.theme.SpaceTiny
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.util.TestTags
+import jp.mydns.kokoichi0206.settings.TestTags
 
 @Composable
 fun ReportIssueScreen(
