+++ feature/settings/src/main/java/jp/mydns/kokoichi0206/settings/pages/ShareAppScreen.kt
@@ -1,4 +1,4 @@
-package jp.mydns.kokoichi0206.sakamichiapp.presentation.setting.pages
+package jp.mydns.kokoichi0206.settings.pages
 
 import android.app.Activity
 import android.content.Intent
@@ -11,7 +11,7 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import jp.mydns.kokoichi0206.sakamichiapp.R
+import jp.mydns.kokoichi0206.feature.settings.R
 import jp.mydns.kokoichi0206.common.ui.theme.SpaceMedium
 import androidx.compose.foundation.background
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -23,10 +23,10 @@ import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.text.font.FontWeight
 import androidx.navigation.NavHostController
 import jp.mydns.kokoichi0206.common.Constants
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.setting.SettingsUiState
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.setting.components.SettingTopBar
+import jp.mydns.kokoichi0206.settings.SettingsUiState
+import jp.mydns.kokoichi0206.settings.components.SettingTopBar
 import jp.mydns.kokoichi0206.common.ui.theme.SpaceSmall
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.util.TestTags
+import jp.mydns.kokoichi0206.settings.TestTags
 
 /**
  * Screen to share this application.
