+++ feature/settings/src/main/res/drawable-v24/app_icon.png
