+++ feature/settings/src/main/res/drawable/link_to_apk.png
