+++ settings.gradle
@@ -16,3 +16,4 @@ include ':feature:member_list'
 include ':feature:member_detail'
 include ':feature:positions'
 include ':feature:quiz'
+include ':feature:settings'
